Migrate ColorGrading to SimpleViewer.
